Add support to change difficulty to render sprites in different modes

+++ Dialog/DoorConfigDialog.cpp
@@ -82,9 +82,9 @@ DoorConfigDialog::DoorConfigDialog(QWidget *parent, LevelComponents::Room *curre
     RenderGraphicsView_Preview();
 
     // Initialize the EntitySet ComboBox
-    for (int i = 0; i < sizeof(entitiessets) / sizeof(entitiessets[0]); ++i)
+    for (unsigned int i = 0; i < sizeof(entitiessets) / sizeof(entitiessets[0]); ++i)
     {
-        comboboxEntitySet.push_back({i, true});
+        comboboxEntitySet.push_back({(int) i, true});
     }
     UpdateComboBoxEntitySet();
